feat: custom fields can be set by plugins for all library items
This is probably the most egregiously disgusting commit, so for that, I apologize. It turns out that creating custom fields required or prompted *a lot* of code changes in order to integrate properly. Thus, summing this commit up as a single "feature" seems laughable, but I couldn't be bothered to properly split it up. I'll try to summarize all the main changes though:

1. Use json in sqlite to allow any plugin to set custom fields. These fields can be queried, edited, accessed, etc. transparently from other fields.
2. Musicbrainz ids were moved to custom fields. This is because I want core fields to solely be anything that is fundamental to a track, album, or extra. Custom fields are for any plugin-specific fields, like in this case, musicbrainz ids.
3. Add custom duplicate detection for items in the library. Because we previously used musibrainz's release-ids for duplicate detection, I added the ability for any plugin to add constraints to whether or not a library item was considered "unique" from another.
4. New duplicate detection and resolution. Duplicate detection, now based off of core and custom constraints, happens automatically whenever an item is changed or added to the library. This functionality, along with the existing duplicate prompt, was moved into a separate 'duplicate' plugin.

I'm sure there's even more that happened here that I forgot about.

Moving musicbrainz ids to custom fields prompted a full database schema rewrite rather than a migration, so this commit is not compatible with any before it. I'm hoping this will be the last database breaking change, and that we can rely on migrations going forward to ensure backwards compatability from this point on.

+++ moe/plugins/remove/rm_core.py
@@ -2,41 +2,42 @@
 
 import logging
 
-import pluggy
+import sqlalchemy
+import sqlalchemy.exc
+from sqlalchemy.orm.session import Session
 
-import moe
 from moe.config import Config, MoeSession
+from moe.library.extra import Extra
 from moe.library.lib_item import LibItem
+from moe.library.track import Track
 
 __all__ = ["remove_item"]
 
 log = logging.getLogger("moe.remove")
 
 
-class Hooks:
-    """Remove plugin hook specifications."""
-
-    @staticmethod
-    @moe.hookspec
-    def post_remove(config: Config, item: LibItem):
-        """Provides an item after it has been removed from the library."""
-
-
-@moe.hookimpl
-def add_hooks(plugin_manager: pluggy.manager.PluginManager):
-    """Registers `add` hookspecs to Moe."""
-    from moe.plugins.remove.rm_core import Hooks
-
-    plugin_manager.add_hookspecs(Hooks)
-
-
 def remove_item(config: Config, item: LibItem):
     """Removes an item from the library."""
     log.debug(f"Removing item from the library. [{item=!r}]")
-
     session = MoeSession()
-    session.delete(item)
+
+    insp = sqlalchemy.inspect(item)
+    if insp.persistent:
+        session.delete(item)
+    elif insp.pending:
+        session.expunge(item)
+        if isinstance(item, (Track, Extra)):
+            item.album_obj = None  # type: ignore
+
+    try:
+        session.flush()
+    except sqlalchemy.exc.InvalidRequestError:
+        # session is currently flushing - delete in separate session to ensure it
+        # occurs before any inserts or updates in the original session
+        if insp.persistent:
+            session.expunge(item)
+            new_session = Session(session.connection())
+            new_session.delete(item)
+            new_session.flush()
 
     log.info(f"Removed item from the library. [{item=!r}]")
-
-    config.plugin_manager.hook.post_remove(config=config, item=item)
